requirements.txt
glfw==2.8.0
numpy==2.2.2
pygame==2.6.1
pyglm==2.8.0
Pygments==2.19.1
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9